LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/stretch/stretch-inline-size-001-expected.txt

CONSOLE MESSAGE: No valid data-* attributes found in selector list : [data-expected-width]








PASS [data-expected-width] 1
FAIL [data-expected-width] 2 assert_equals:
<canvas class="test" data-expected-width="45"></canvas>
width expected 45 but got 50
FAIL [data-expected-width] 3 assert_equals:
<input class="test" data-expected-width="45">
width expected 45 but got 152
FAIL [data-expected-width] 4 assert_equals:
<textarea class="test" data-expected-width="45"></textarea>
width expected 45 but got 165
FAIL [data-expected-width] 5 assert_equals:
<button class="test" data-expected-width="45"></button>
width expected 45 but got 10
FAIL [data-expected-width] 6 assert_equals:
<div class="test" style="float: left" data-expected-width="45"></div>
width expected 45 but got 10
PASS [data-expected-width] 7
FAIL [data-expected-width] 8 assert_equals:
<div class="test" style="position: absolute" data-expected-width="50"></div>
width expected 50 but got 10
FAIL [data-expected-width] 9 assert_equals:
<div class="test" style="position: absolute; inset-inline-start: 10px" data-expected-width="45"></div>
width expected 45 but got 10
FAIL [data-expected-width] 10 assert_equals:
<div class="test" style="position: absolute; inset-inline-end: 10px" data-expected-width="45"></div>
width expected 45 but got 10
PASS [data-expected-width] 11
PASS [data-expected-width] 12
PASS [data-expected-width] 13
FAIL [data-expected-width] 14 assert_equals:
<div class="test" style="position: absolute" data-expected-width="55"></div>
width expected 55 but got 10
FAIL [data-expected-width] 15 assert_equals:
<div class="test" style="position: absolute; inset-inline-start: 10px" data-expected-width="45"></div>
width expected 45 but got 10
FAIL [data-expected-width] 16 assert_equals:
<div class="test" style="position: absolute; inset-inline-end: 10px" data-expected-width="45"></div>
width expected 45 but got 10
PASS [data-expected-width] 17
PASS [data-expected-width] 18
FAIL [data-expected-width] 19 assert_equals:
<div class="test" data-expected-width="45"></div>
width expected 45 but got 10
PASS [data-expected-width] 20
PASS [data-expected-width] 21
PASS [data-expected-width] 22
PASS [data-expected-width] 23
